+++ W040422MR - 2022.11.21/real-app-front/src/App.jsx
@@ -4,6 +4,8 @@ import Navbar from "./components/navbar";
 import Home from "./components/home";
 import About from "./components/about";
 
+import { Route, Routes } from "react-router-dom";
+
 function App() {
   return (
     <div className="app d-flex flex-column min-vh-100">
@@ -12,8 +14,10 @@ function App() {
       </header>
 
       <main className="flex-fill container">
-        {/* <Home /> */}
-        <About />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="about" element={<About />} />
+        </Routes>
       </main>
 
       <Footer />