+++ W040422MR - 2022.11.21/real-app-front/src/components/navbar.jsx
@@ -1,10 +1,12 @@
+import { Link, NavLink } from "react-router-dom";
+
 const Navbar = () => {
   return (
     <nav className="navbar navbar-expand-sm navbar-light bg-light shadow-sm">
       <div className="container-fluid">
-        <a className="navbar-brand" href="#">
+        <Link to="/" className="navbar-brand">
           Real<i className="bi bi-geo-fill"></i>App
-        </a>
+        </Link>
         <button
           className="navbar-toggler"
           type="button"
@@ -20,27 +22,27 @@ const Navbar = () => {
         <div className="collapse navbar-collapse" id="navbarsExample03">
           <ul className="navbar-nav me-auto mb-2 mb-sm-0">
             <li className="nav-item">
-              <a className="nav-link" href="#">
+              <NavLink to="about" className="nav-link">
                 About
-              </a>
+              </NavLink>
             </li>
             <li className="nav-item">
-              <a className="nav-link" href="#">
+              <NavLink to="my-cards" className="nav-link">
                 My Cards
-              </a>
+              </NavLink>
             </li>
           </ul>
 
           <ul className="navbar-nav ms-auto mb-2 mb-sm-0">
             <li className="nav-item">
-              <a className="nav-link" href="#">
+              <NavLink to="sign-in" className="nav-link">
                 Sign In
-              </a>
+              </NavLink>
             </li>
             <li className="nav-item">
-              <a className="nav-link" href="#">
+              <NavLink to="sign-up" className="nav-link">
                 Sign Up
-              </a>
+              </NavLink>
             </li>
           </ul>
         </div>